strip-harbor/requirements.txt
django
stripe
django-environ
mailersend